Move write_tool to separate file
